Hide `lightyear_frame_interpolation` tests behind unused module
